+++ tests/api-resources/realtime.test.ts
@@ -1,0 +1,23 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+import Micro from '@micro-so/sdk';
+
+const client = new Micro({
+  apiKey: 'My API Key',
+  teamID: 'My Team ID',
+  baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
+});
+
+describe('resource realtime', () => {
+  // Mock server tests are disabled
+  test.skip('createTicket', async () => {
+    const responsePromise = client.realtime.createTicket();
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+});